% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{refutation_sampler}
\alias{refutation_sampler}
\title{Refutation estimation function}
\usage{
refutation_sampler(
  effect_size = c("RR", "OR"),
  current_RCT_mean,
  current_RCT_lb,
  current_RCT_ub,
  event_percentage,
  cutoff,
  total_future_patients
)
}
\arguments{
\item{effect_size}{Effect size type, "RR" or "OR"}

\item{current_RCT_mean}{Mean effect for the current study}

\item{current_RCT_lb}{Lower bound for the confidence interval for the effect from the current study}

\item{current_RCT_ub}{Upper bound for the confidence interval for the effect from the current study}

\item{event_percentage}{Base rate/risk for the neutral hypothetical study}

\item{cutoff}{Lower limit for ROPE}

\item{total_future_patients}{Size of hypothetica neutral future trial}
}
\value{

}
\description{
Refutation estimation function
}
\examples{
refutation_sampler(effect_size = "RR" ,  # effect_size
current_RCT_mean = 0.7 ,                    # current_RCT_mean
current_RCT_lb = 0.51 ,                     # current_RCT_lb
current_RCT_ub = 0.96 ,                     # current_RCT_ub
event_percentage = 45 ,               # event_percentage
cutoff = 0.9 )                     # lower limit for ROPE
}
